+++ src/Octoshift/RetryPolicy.cs
@@ -29,7 +29,7 @@ namespace OctoshiftCLI
             return await policy.ExecuteAsync(func);
         }
 
-        public async Task<PolicyResult<T>> RetryOnResult<T>(Func<Task<T>> func, T resultFilter, string retryLogMessage = null)
+        public async Task<PolicyResult<T>> RetryOnResult<T>(Func<Task<T>> func, Func<T, bool> resultFilter, string retryLogMessage = null)
         {
             var policy = Policy.HandleResult(resultFilter)
                                .WaitAndRetryAsync(5, retry => retry * TimeSpan.FromMilliseconds(_retryInterval), (_, _) =>